requirements.txt
easydict==1.6
greenlet==0.4.10
motor==0.2
motorengine==0.9.0
pymongo==2.7
six==1.10.0
sockjs-tornado==1.0.3
tornado==4.4.2
